Lift TransitionsEmitter.PhaseGraph to a shared internal type.
The termination-reachability guard and the emitted ValidTransitions table
must resolve successors from one graph so they cannot drift.
